+++ packages/taroize/src/wxml.ts
@@ -2,12 +2,10 @@ import { parse } from 'himalaya-wxml'
 import * as t from 'babel-types'
 import { camelCase, cloneDeep } from 'lodash'
 import traverse, { NodePath, Visitor } from 'babel-traverse'
-import { buildTemplate, DEFAULT_Component_SET, buildImportStatement, buildBlockElement } from './utils'
+import { buildTemplate, DEFAULT_Component_SET, buildImportStatement, buildBlockElement, parseCode } from './utils'
 import { specialEvents } from './events'
 import { parseTemplate, parseModule } from './template'
 import { usedComponents, errors } from './global'
-import * as fs from 'fs'
-import { resolve } from 'path'
 import { reserveKeyWords } from './constant'
 import { parseExpression } from 'babylon'
 
@@ -61,9 +59,10 @@ type AttrValue =
   | t.JSXExpressionContainer
   | null
 
-interface Imports {
+export interface Imports {
   ast: t.File,
-  name: string
+  name: string,
+  wxs?: boolean
 }
 
 const WX_IF = 'wx:if'
@@ -72,6 +71,16 @@ const WX_FOR = 'wx:for'
 const WX_FOR_ITEM = 'wx:for-item'
 const WX_FOR_INDEX = 'wx:for-index'
 const WX_KEY = 'wx:key'
+
+export const wxTemplateCommand = [
+  WX_IF,
+  WX_ELSE_IF,
+  WX_FOR,
+  WX_FOR_ITEM,
+  WX_FOR_INDEX,
+  WX_KEY,
+  'wx:else'
+]
 
 function buildElement (
   name: string,
@@ -91,8 +100,7 @@ export const createWxmlVistor = (
   refIds: Set<string>,
   dirPath: string,
   wxses: WXS[] = [],
-  imports: Imports[] = [],
-  className?: string
+  imports: Imports[] = []
 ) => {
   return {
     JSXAttribute (path) {
@@ -112,8 +120,18 @@ export const createWxmlVistor = (
         }
       }
     },
-    BlockStatement () {
-      // debugger
+    JSXIdentifier (path) {
+      const nodeName = path.node.name
+      if (path.parentPath.isJSXAttribute()) {
+        if (nodeName === WX_KEY) {
+          path.replaceWith(t.jSXIdentifier('key'))
+        }
+        if (nodeName.startsWith('wx:') && !wxTemplateCommand.includes(nodeName)) {
+          // tslint:disable-next-line
+          console.log(`未知 wx 作用域属性： ${nodeName}，该属性会被移除掉。`)
+          path.parentPath.remove()
+        }
+      }
     },
     JSXElement: {
       enter (path: NodePath<t.JSXElement>) {
@@ -148,7 +166,7 @@ export const createWxmlVistor = (
           }
         }
         if (tagName === 'Wxs') {
-          wxses.push(getWXS(attrs.map(a => a.node), path, dirPath))
+          wxses.push(getWXS(attrs.map(a => a.node), path, imports))
         }
         if (tagName === 'Template') {
           const template = parseTemplate(path, dirPath)
@@ -181,7 +199,7 @@ export const createWxmlVistor = (
               }
             })
             usedTemplate.forEach(componentName => {
-              if (componentName !== className) {
+              if (componentName !== classDecl.id.name) {
                 ast.program.body.unshift(
                   buildImportStatement(`./${componentName}`, [], componentName)
                 )
@@ -235,6 +253,7 @@ export function parseWXML (dirPath: string, wxml?: string, parseImport?: boolean
     errors.length = 0
   }
   usedComponents.clear()
+  usedComponents.add('Block')
   let wxses: WXS[] = []
   let imports: Imports[] = []
   const refIds = new Set<string>()
@@ -262,7 +281,7 @@ export function parseWXML (dirPath: string, wxml?: string, parseImport?: boolean
   traverse(ast, createWxmlVistor(loopIds, refIds, dirPath, wxses, imports))
 
   refIds.forEach(id => {
-    if (loopIds.has(id)) {
+    if (loopIds.has(id) || imports.filter(i => i.wxs).map(i => i.name).includes(id)) {
       refIds.delete(id)
     }
   })
@@ -275,7 +294,7 @@ export function parseWXML (dirPath: string, wxml?: string, parseImport?: boolean
   }
 }
 
-function getWXS (attrs: t.JSXAttribute[], path: NodePath<t.JSXElement>, dirPath: string): WXS {
+function getWXS (attrs: t.JSXAttribute[], path: NodePath<t.JSXElement>, imports: Imports[]): WXS {
   let moduleName: string | null = null
   let src: string | null = null
 
@@ -310,7 +329,11 @@ function getWXS (attrs: t.JSXAttribute[], path: NodePath<t.JSXElement>, dirPath:
       throw new Error('wxs 如果没有 src 属性，标签内部必须有 wxs 代码。')
     }
     src = './wxs__' + moduleName
-    fs.writeFileSync(resolve(dirPath, src + '.wxs'), script.value)
+    imports.push({
+      ast: parseCode(script.value),
+      name: moduleName as string,
+      wxs: true
+    })
   }
 
   if (!moduleName || !src) {
@@ -379,9 +402,6 @@ function transformLoop (
           index = node.value
           p.remove()
         }
-        if (node.name.name === WX_KEY) {
-          p.get('name').replaceWith(t.jSXIdentifier('key'))
-        }
       })
 
     const replacement = t.jSXExpressionContainer(
@@ -423,7 +443,7 @@ function transformIf (
   const conditions: Condition[] = []
   let siblings: NodePath<t.Node>[] = []
   try {
-    siblings = jsx.getAllNextSiblings()
+    siblings = jsx.getAllNextSiblings().filter(s => !(s.isJSXExpressionContainer() && s.get('expression').isJSXEmptyExpression()))
   } catch (error) {
     return
   }
@@ -582,6 +602,7 @@ function removEmptyTextAndComment (nodes: AllKindNode[]) {
   return nodes.filter(node => {
     return node.type === NodeType.Element
       || (node.type === NodeType.Text && node.content.trim().length !== 0)
+      || node.type === NodeType.Comment
   })
 }
 
@@ -687,11 +708,11 @@ function handleAttrKey (key: string) {
     return key
   } else if (key === 'class') {
     return 'className'
-  } else if (/^(bind|catch)[a-z]/.test(key)) {
+  } else if (/^(bind|catch)[a-z|:]/.test(key)) {
     if (specialEvents.has(key)) {
       return specialEvents.get(key)!
     } else {
-      key = key.replace(/^(bind|catch)[a-z]/, 'on')
+      key = key.replace(/^(bind:|catch:|bind|catch)/, 'on')
       return key.substr(0, 2) + key[2].toUpperCase() + key.substr(3)
     }
   }